New PiInterpreter API to map logical port to long type
Introduce a new API for PiInterpreter to map logical port number to a long
port number and deprecate the old one.
The reason we need to translate logical port number to a long type is that
the mapped number will become negative when translating port number to
an unsigned integer with MSB (e.g., 0xFFFFFFFD) and the low level driver
or translator is expecting to get a positive number.

+++ pipelines/fabric/impl/src/main/java/org/onosproject/pipelines/fabric/impl/behaviour/FabricCapabilities.java
@@ -49,7 +49,7 @@ public class FabricCapabilities {
                 .table(FabricConstants.FABRIC_INGRESS_NEXT_HASHED).isPresent();
     }
 
-    public Optional<Integer> cpuPort() {
+    public Optional<Long> cpuPort() {
         // This is probably brittle, but needed to dynamically get the CPU port
         // for different platforms.
         if (!pipeconf.extension(CPU_PORT_TXT).isPresent()) {
@@ -67,7 +67,7 @@ public class FabricCapabilities {
                 return Optional.empty();
             }
             try {
-                return Optional.of(Integer.parseInt(str));
+                return Optional.of(Long.parseLong(str));
             } catch (NumberFormatException e) {
                 log.error("Invalid CPU port for {}: {}", pipeconf.id(), str);
                 return Optional.empty();
